surya and all- changes merged

+++ src/app/districtadmin-menu/appoint-district-off-bearer/appoint-district-off-bearer.component.ts
@@ -14,10 +14,14 @@ export class AppointDistrictOffBearerComponent implements OnInit {
   router: any;
   editform:FormGroup;
   hidden:boolean=true;
+  district:string;
+  constituency:string='';
+  
+
 
   constructor(public ApiService:ApiServiceService,
-    private fb: FormBuilder) 
-    { 
+    private fb: FormBuilder)
+    {
       this.officebearerform = this.fb.group({ //angForm
         email: ['', [Validators.required, Validators.email,Validators.pattern('^[a-z0-9._%+-]+@[a-z0-9.-]+\\.[a-z]{2,4}$')]],
         firstname:['',[Validators.required, Validators.pattern('[A-Za-z ]{1,32}')]],
@@ -70,7 +74,7 @@ export class AppointDistrictOffBearerComponent implements OnInit {
   customers:any=[];
   officebearerform !:FormGroup;
   dtOptions: DataTables.Settings = {};
-  
+
   ngOnInit(): void {
     // this.getdata();
     // this.ApiService.viewtableOB();
@@ -86,8 +90,21 @@ export class AppointDistrictOffBearerComponent implements OnInit {
     this.dtOptions = {
       pagingType: 'full_numbers'
     };
-    
-   
+
+
+  }
+  district_list:any[]=this.ApiService.all_districts;
+  constituency_list:any=this.ApiService.all_constituency;
+  user_constituency:any;
+  Constituency_selection(selection:any){
+    //console.log(selection)
+    if(selection != 'null'){
+      //console.log(this.district);
+      let obj=this.constituency_list;
+      this.user_constituency=obj[this.district];
+
+    }
+
   }
   getdata(){
     this.customers=[];
@@ -124,27 +141,30 @@ export class AppointDistrictOffBearerComponent implements OnInit {
 
   }
   delete_ob(user_id : any)
-    {
-        console.log(user_id)
-            this.ApiService.delete_admin(user_id)
-            .pipe()
-            .subscribe(
-            data => {
-                window.location.reload();
-                //this.router.navigate(['uikit/formlayout']);
-                alert("Office Bearer detail has been deleted !")
-            },
+  {
+      console.log(user_id)
+      if(confirm("Are you sure want to delete this record ?")) {
+        console.log("Implement delete functionality here");
+          this.ApiService.delete_admin(user_id)
+          .pipe()
+          .subscribe(
+          data => {
+              window.location.reload();
+              //this.router.navigate(['uikit/formlayout']);
+              alert("Office Bearer detail has been deleted !")
+          },
 
-            error => {
-                console.log(error);
-            });
+          error => {
+              console.log(error);
+          });
 
-    }
+  }
+}
     get email() { return this.officebearerform.get('email'); }
     get firstname() { return this.officebearerform.get('firstname'); }
     get lastname() { return this.officebearerform.get('lastname'); }
     get applied_role() { return this.officebearerform.get('applied_role'); }
-  
+
 
     OBid:any;
     OBname: any;
@@ -186,13 +206,13 @@ export class AppointDistrictOffBearerComponent implements OnInit {
          this.OBaddress=a.address1;
          this.OBold_designation=a.applied_role;
          this.OBcomments=a.party_comments;
-         
+
         this.fullname1=a.name;
 
          this.reqform.patchValue({
           id1:this.OBid,
           email1:this.OBmail,
-          name:this.fullname1,       
+          name:this.fullname1,
           old_designation:this.OBold_designation,
           new_designation1:"",
           reason:""
@@ -262,7 +282,7 @@ postdata1(angForm1) //angForm1
         data => {
           // console.log(angForm1.value.name,angForm1.value.user_id,angForm1.value.new_designation,angForm1.value.old_designation,angForm1.value.reason );
             alert("Request has been created successfully!")
-   
+
         this.router.navigate(['superadmin/Approve-Reject']);
         angForm1.reset();
         },
@@ -276,6 +296,22 @@ postdata1(angForm1) //angForm1
     }
 }
 
+
+date_of_birth :string;
+age : number;
+calculateAge() {
+  console.log(this.date_of_birth);
+  console.log("i m in");
+
+  const today = new Date();
+  const birthdate = new Date(this.date_of_birth);
+  this.age = today.getFullYear() - birthdate.getFullYear();
+  const m = today.getMonth() - birthdate.getMonth();
+  if (m < 0 || (m === 0 && today.getDate() < birthdate.getDate())) {
+    this.age--;
+  }
+
+}
